requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
streamlit>=1.28.0
langchain-groq>=0.1.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
fpdf2>=2.7.0
pandas>=2.0.0
Flask>=3.0.0
gunicorn>=23.0.0
flask-cors>=4.0.0
pymongo>=4.8.0
Authlib>=1.3.1
